Separate Home and Editor
