+++ src/App.tsx
@@ -1,13 +1,13 @@
 import { AuthProvider, useAuth } from "./context/AuthContext";
 
 import Auth from "./pages/Auth";
-import Editor from "./pages/Editor";
+import Home from "./pages/Home";
 
 function Inner() {
   const { user } = useAuth();
 
   if (user) {
-    return <Editor user={user} />;
+    return <Home user={user} />;
   }
   return <Auth />;
 }
